[REFACTOR] FAQRepository, NoticeRepository 패키지 변경
