+++ src/main/java/ject/componote/domain/announcement/dao/FAQRepository.java
@@ -1,5 +1,6 @@
-package ject.componote.domain.announcement.domain;
+package ject.componote.domain.announcement.dao;
 
+import ject.componote.domain.announcement.domain.FAQ;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FAQRepository extends JpaRepository<FAQ, Long> {
