+++ src/main/java/ject/componote/domain/announcement/dao/NoticeRepository.java
@@ -1,5 +1,6 @@
-package ject.componote.domain.announcement.domain;
+package ject.componote.domain.announcement.dao;
 
+import ject.componote.domain.announcement.domain.Notice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface NoticeRepository extends JpaRepository<Notice, Long> {
